+++ SONIC_JK/Act1_2_sky2.cpp
@@ -31,8 +31,7 @@ namespace jk
 		Transform* tr = GetComponent<Transform>();
 		Vector2 pos = tr->GetPos();
 		Vector2 mpos = Camera::CaluatePos(pos);
-		TransparentBlt(hdc, static_cast<int>(mpos.x), static_cast<int>(mpos.y), 14064, 897,
-			mImage->GetHdc(), 0, 0, 14064, 897, RGB(0, 0, 0));
+		TransparentBlt(hdc, mpos.x, mpos.y, 14064, 897, mImage->GetHdc(), 0, 0, 14064, 897, RGB(0, 0, 0));
 		Gameobject::Render(hdc);
 	}
