+++ SONIC_JK/Act1_2_sky3.cpp
@@ -29,8 +29,7 @@ namespace jk
 		Vector2 pos = tr->GetPos();
 		tr->SetPos(pos);
 		Vector2 mpos = Camera::CaluatePos(pos);
-		TransparentBlt(hdc, static_cast<int>(mpos.x / 3.3f), static_cast<int>(mpos.y), 14064, 240,
-			mImage->GetHdc(), 0, 0, 14064, 240, RGB(13, 72, 7));
+		TransparentBlt(hdc, mpos.x / 3.3f, mpos.y, 14064, 240, mImage->GetHdc(), 0, 0, 14064, 240, RGB(13, 72, 7));
 
 		Gameobject::Render(hdc);
 	}
