+++ SONIC_JK/Ending_boss.cpp
@@ -35,7 +35,7 @@ namespace jk
 	{
 		Transform* tr = GetComponent<Transform>();
 		pos =tr->GetPos();		
-		pos.y += 150 * static_cast<float>(Time::DeltaTime());
+		pos.y += 150 * Time::DeltaTime();
 		tr->SetPos(pos);
 		Gameobject::Update();
 	}
