+++ SONIC_JK/jk_Act1_2_BG.cpp
@@ -29,8 +29,7 @@ namespace jk
 		Transform* tr = GetComponent<Transform>();	
 		Vector2 pos = tr->GetPos();
 		Vector2 mpos = pos - (Camera::CaluatePos(pos) - pos);
-		TransparentBlt(hdc, 0, 0, 1200, 840, 
-			mImage->GetHdc(), static_cast<int>(mpos.x), static_cast<int>(mpos.y), 1200, 840, RGB(4, 20, 83));
+		TransparentBlt(hdc, 0, 0, 1200, 840, mImage->GetHdc(), mpos.x, mpos.y, 1200, 840, RGB(4, 20, 83));
 		Gameobject::Render(hdc);
 	}
 	void Act1_2_BG::Release()
