+++ SONIC_JK/jk_Act1_Water.cpp
@@ -30,8 +30,8 @@ namespace jk
 		Vector2 pos = tr->GetPos();
 		tr->SetPos(Vector2{ 21240.f,  3936.f });
 
-		float pos_y = mImage->GetHeight() * 3.5f;
-		float pos_x = mImage->GetWidth() * 3.f;
+		float pos_y = mImage->GetHeight() * 3.5;
+		float pos_x = mImage->GetWidth() * 3;
 
 		Collider* collider = AddComponent<Collider>();
 		collider->SetSize(Vector2(pos_x, pos_y));		
@@ -57,8 +57,7 @@ namespace jk
 		func.SourceConstantAlpha = (BYTE)checktime;
 		Vector2 mpos = Camera::CaluatePos(pos);
 		
-		AlphaBlend(hdc, static_cast<int>(mpos.x), static_cast<int>(mpos.y), mImage->GetWidth()*3, mImage->GetHeight()* static_cast<int>(3.5f),
-			mImage->GetHdc(), 0, 0, mImage->GetWidth(), mImage->GetHeight(), func);
+		AlphaBlend(hdc,  mpos.x, mpos.y, mImage->GetWidth()*3, mImage->GetHeight()*3.5, mImage->GetHdc(), 0, 0, mImage->GetWidth(), mImage->GetHeight(), func);
 	
 		Gameobject::Render(hdc);
 	}
