+++ SONIC_JK/jk_Collider.cpp
@@ -63,7 +63,7 @@ namespace jk
 
 			if (ONOFF == 0)
 			{
-				Rectangle(hdc, static_cast<int>(pos.x), static_cast<int>(pos.y), static_cast<int>(pos.x + mSize.x), static_cast<int>(pos.y + mSize.y));
+				Rectangle(hdc, pos.x, pos.y, pos.x + mSize.x, pos.y + mSize.y);
 			}
 			//Rectangle(hdc, pos.x - mSize.x/2, pos.y - mSize.y, pos.x + mSize.x /2, pos.y);
